Refactor: Album 페이지폴더로 변경
