+++ lib/main.dart
@@ -8,7 +8,7 @@ import 'package:krosscutting_app/provider/video_path_provider.dart';
 import 'package:krosscutting_app/provider/video_direction_provider.dart';
 
 import 'package:krosscutting_app/screens/select_screen/video_manager.dart';
-import 'package:krosscutting_app/screens/album_screen.dart';
+import 'package:krosscutting_app/screens/album_screen/album_screen.dart';
 import 'package:krosscutting_app/screens/home_screen/home_screen.dart';
 import 'package:krosscutting_app/screens/splash_screen.dart';
 import 'package:krosscutting_app/screens/select_screen/video_select_start_point.dart';
