+++ lib/screens/album_screen/embedded_album.dart
@@ -1,7 +1,6 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:image_picker/image_picker.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 
@@ -9,17 +8,18 @@ import 'package:krosscutting_app/constants/type.dart';
 import 'package:krosscutting_app/provider/video_path_provider.dart';
 import 'package:krosscutting_app/widgets/gradient_button.dart';
 import 'package:krosscutting_app/screens/select_screen/video_manager.dart';
+import 'package:krosscutting_app/screens/album_screen/selection_row.dart';
 
-class ImbeddedAlbum extends StatefulWidget {
-  const ImbeddedAlbum({
+class EmbeddedAlbum extends StatefulWidget {
+  const EmbeddedAlbum({
     super.key,
   });
 
   @override
-  State<ImbeddedAlbum> createState() => _ImbeddedAlbumState();
+  State<EmbeddedAlbum> createState() => _EmbeddedAlbumState();
 }
 
-class _ImbeddedAlbumState extends State<ImbeddedAlbum> {
+class _EmbeddedAlbumState extends State<EmbeddedAlbum> {
   bool isAllSelected = false;
 
   @override
@@ -28,6 +28,7 @@ class _ImbeddedAlbumState extends State<ImbeddedAlbum> {
   }
 
   void uploadFile(videoPathMap, context) async {
+    //TODO: endPoints | 수정 app/videos/contents/files
     var url = Uri.parse("${dotenv.env["SERVER_HOST"]}/videos/contents/files");
     var request = http.MultipartRequest("POST", url);
 
@@ -51,10 +52,9 @@ class _ImbeddedAlbumState extends State<ImbeddedAlbum> {
 
     if (response.statusCode == 200) {
       updateVideoManager(context);
-
       Navigator.pushNamed(context, "/selection/startpoint");
     } else {
-      //TO DO: 유저에게 안내
+      //TODO: 유저에게 안내
     }
   }
 
@@ -98,98 +98,5 @@ class _ImbeddedAlbumState extends State<ImbeddedAlbum> {
               )
       ],
     );
-  }
-}
-
-class SelectionRow extends StatefulWidget {
-  final Map type;
-
-  const SelectionRow({
-    super.key,
-    required this.type,
-  });
-
-  @override
-  State<SelectionRow> createState() => _SelectionRowState();
-}
-
-class _SelectionRowState extends State<SelectionRow> {
-  @override
-  Widget build(BuildContext context) {
-    Size screenSize = MediaQuery.of(context).size;
-
-    return SizedBox(
-      height: screenSize.height * 0.13,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Text(
-            widget.type["buttonName"],
-            style: TextStyle(
-              fontSize: screenSize.height * 0.04,
-            ),
-          ),
-          SelectButton(
-            typeKey: widget.type["key"],
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class SelectButton extends StatefulWidget {
-  final String typeKey;
-
-  const SelectButton({
-    super.key,
-    required this.typeKey,
-  });
-
-  @override
-  State<SelectButton> createState() => _SelectButtonState();
-}
-
-class _SelectButtonState extends State<SelectButton> {
-  @override
-  Widget build(BuildContext context) {
-    final videoPathProvider = Provider.of<VideoPathProvider>(context);
-    final videoPathMap = videoPathProvider.videoPath;
-    bool isSelected = videoPathMap[widget.typeKey] != null;
-
-    Future<void> pickFiles() async {
-      XFile? selectedVideo =
-          await ImagePicker().pickVideo(source: ImageSource.gallery);
-
-      if (selectedVideo != null) {
-        videoPathProvider.setVideoPath(widget.typeKey, selectedVideo.path);
-      }
-    }
-
-    return isSelected
-        ? Row(
-            children: [
-              const Icon(
-                Icons.check_circle,
-                size: 40,
-                color: Colors.deepPurple,
-              ),
-              IconButton(
-                onPressed: pickFiles,
-                icon: const Icon(
-                  Icons.replay_circle_filled_rounded,
-                  size: 40,
-                  color: Color.fromARGB(255, 30, 233, 209),
-                ),
-              ),
-            ],
-          )
-        : IconButton(
-            icon: const Icon(
-              size: 40,
-              Icons.video_collection_rounded,
-            ),
-            onPressed: pickFiles,
-          );
   }
 }
